+++ src/test/kotlin/dev/ekvedaras/laravelquery/inspection/DasObjectsInspectionTest.kt
@@ -94,4 +94,8 @@ internal class DasObjectsInspectionTest : BaseTestCase() {
     fun testDoesNotWarnAboutKnownJsonColumn() {
         assertInspection("inspection/knownJsonColumn.php", UnknownColumnInspection())
     }
+
+    fun testItDoesNotWarnAboutUnknownColumnInObjectsAsCreateFunctionValues() {
+        assertInspection("inspection/newObjectWithinCreateValue.php", UnknownColumnInspection())
+    }
 }