+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Laravel Query Changelog
 
 ## [Unreleased]
+### Fixed
+- Do not inspect for unknown column in case its value of object. #96
 
 ## [3.3.0]
 ### Added